init other tests and modularize #166
