+++ tests/sanity/test_sanityValidation.py
@@ -1,0 +1,4 @@
+from feelpp.benchmarking.reframe.config.configSchemas import Sanity
+
+class TestSanity:
+    pass
